Renamed get_home_dir to get_common_dirs
