+++ src/commands/purge.rs
@@ -12,7 +12,7 @@ use crate::{
     containers::cleansweep_file_paths::CleansweepFilePaths,
     systems::json_io::read_file_to_struct,
     utils::{
-        file_to_string_vec::file_to_string_vec, get_home_dir::get_cleansweep_dir,
+        file_to_string_vec::file_to_string_vec, get_common_dirs::get_cleansweep_dir,
         run_time_user_input::get_string_input_matching_provided_string,
     },
 };
