+++ src/commands/setup.rs
@@ -6,7 +6,7 @@ use thiserror::Error;
 use crate::containers::cleansweep_file_paths::CleansweepFilePaths;
 use crate::containers::user_settings::UserSettings;
 use crate::utils::create_defaults::create_default_user_settings;
-use crate::utils::get_home_dir::get_cleansweep_dir;
+use crate::utils::get_common_dirs::get_cleansweep_dir;
 
 use crate::systems::json_io::*;
