+++ src/utils/get_common_dirs.rs
