+++ src/utils/mod.rs
@@ -1,5 +1,5 @@
 pub mod create_defaults;
 pub mod file_to_string_vec;
-pub mod get_home_dir;
+pub mod get_common_dirs;
 pub mod path_types_to_string;
 pub mod run_time_user_input;
